core: Create 16 PCI-E ports for Q35
Currently we have a single unused pcie-root-port available for
hotplugging devices. When we are out of port we cannot add devices
anymore. In order to to avoid this, create 16 PCI Express Root Ports on
the default configuration that can be used either by devices present at
domain creation time or for hotplug.

Change-Id: I6d5c422b8bc37c322dc14dfb5d063f66ed8e813e
Bug-Url: https://bugzilla.redhat.com/1527882
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>