engine: setup single ipv6 gateway for default route role
With this patch the meaning of the default route role changes.
It now signifies a network which provides either v4 or v6 default routes
on a host (or both).

When passing multiple ipv6 gateways to vdsm in setup networks request:
 - vdsm adds more than one to the routing table (instead of replacing the
   default route as is done for ipv4)
 - vdsm reports back all gateways as 'unspecified' ('::') because there is
   more than one gateway.

Therefore in engine it is necessary to send only a single ipv6 gateway
to vdsm to enforce a single default route entry in the v6 routing table
and remove the gateway from the previous default route role network if
the role was moved to another network.

In setup networks command:
 - determine whether an attachment was associated with the default route
   role network by checking both for v4 and for v6.
 - remove the ipv6 gateway from the attachment of the previous default
   route role network.
 - add the attachment of the previous default route role network to the
   attachments being re-configured. This guarantees that the ipv6
   gateway of the corresponding interface is reconfigured to 'unspecified'
   in the  next step.
 - persist the modified previous attachment to ensure the attachment
   reflects the state of the interface and is in sync with it.

On the data structure that is sent to vdsm in setup networks:
 - on the interface attached to the default route role network,
   set the actual ipv6 gateway which is configured on it
 - on all interfaces which are not attached to the current default
   route role network - replace the gateway to 'unspecified'

An 'unspecified' gateway is designated as:
 - "::" or null on the attachment
 - "::" on the interface

Change-Id: Ibe952108fdaeb38f9a909ec1b9bab6cead296ae8
Bug-Url: https://bugzilla.redhat.com/1467332
Signed-off-by: Eitan Raviv <eraviv@redhat.com>