restapi: Set the template id to Guid.Empty()
When trying to import a VM from ovf configuration (i.e. create a new
VM from ovf), and in case the original VM was thin provisioned
(dependent of the template), removing it removes the VM disks
permanently instead of detaching them as they are dependent on the
template disks.
Therefore, querying for the template's disks won't be useful.

This patch sets the template's id to Guid.Empty(), so the VM will be
created from the ovf configuration but without any disks.

Change-Id: I3b3557044a5e7a4fa9c7dcc84a2b6bb198084343
Bug-Url: https://bugzilla.redhat.com/1658976
Signed-off-by: Shani Leviim <sleviim@redhat.com>