engine: USB Forwarding fix after Upgrade
When upgrading from a version which did not support the Domain XML
feature to a version that does support the Domain XML feature, the USB
Fowarding failed because the previous version set the plugged field for
the spicevmc devices of type redir to false.

The problem is that when the Domain XML feature is enabled it does use
the spicevmc devices of type redir and therefore the plugged field needs
to be reset in order to enable these devices. This patch copies the
redirdev addresses to the redir devices when available and sets the
is_plugged flag of the redir device to true during setup and during
monitoring.

Change-Id: If5c118479b32ad47feba4d40c2038598e319c42a
Bug-Url: https://bugzilla.redhat.com/1663616
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>