restapi: Fix Disk-Attachment hrefs under Vm
Under Vms, disk-attachments have invalid hrefs due to limitation of
link-generation infrastructure.

The infrastructure identifies a location in the API-tree by the combination:
entity-type, parent-type.

In this particular case, the combination of disk-attachment under a VM exists
twice - once under root, once under (export) Storage-Domain:

.../api/vms/xxx/diskattachments
.../api/storagedomains/zzz/vms/xxx/diskattachments

This patch adds manual creation of the href to overcome this problem

See also the bug and patch for the second location:
https://bugzilla.redhat.com/1374323
https://gerrit.ovirt.org/#/c/93507/

Bug-Url: https://bugzilla.redhat.com/1647018
Change-Id: Ie9a5c819d18bec3edba4a48d7b0afbabcf85523f
Signed-off-by: Ori_Liel <oliel@redhat.com>