engine: MBS check if connector info available
In the Run VM flow, if the VM has Managed Block disks
we need to check if the connector information is available
on the selected host before trying to connect the disks.

The connector info will be available only if the os-brick
package is installed in the host.

If the connector info is not available, an error log is
generated and the Run VM operations fails.

Change-Id: Id7ca6de70330f3fa7eb740e57a197363bed2398f
Bug-Url: https://bugzilla.redhat.com/1670059
Signed-off-by: Fred Rolland <frolland@redhat.com>