restapi: Install/Approve Host - allow providing 'activate' in request body
When installing or activating a host, currently 'acitvate' may be provided
as a URL parameter (see: https://gerrit.ovirt.org/#/c/96107/). This is
somewhat unnatural for 'actions' (install, approve) and so this patch
enables providing it in the request body as well. If both exist the
one in the request body takes precedence.

Bug-Url: https://bugzilla.redhat.com/1561539
Change-Id: I3e9047d5c784330581475d6d2aab745327dbbf32
Signed-off-by: Ori_Liel <oliel@redhat.com>