packaging: dbscripts: Improve logging
Output the various "titles" to the log file, instead of to stdout.

We call psql with --log-file=DBFUNC_LOGFILE, so it logs to there
immediately. But the normal output (to stdout) goes there only by
engine-setup (otopi), after schema.sh finishes, so it's impossible to
match the title with the psql output. So:

Replace all relevant calls to 'echo' with calls to a new function
dbfunc_output that appends to DBFUNC_LOGFILE if set, otherwise to
stdout.

Replace most calls to dbfunc_psql and dbfunc_psql_die with calls to new
functions dbfunc_psql_v and dbfunc_psql_die_v that log/output that they
were called.

Instead of pipes like 'psql|psql', save the output to a variable
and call on it.

Change-Id: I29e558b6b14564a223b5baca763f4e1a24e8e174
Signed-off-by: Yedidyah Bar David <didi@redhat.com>