core: log GetVmLeaseInfo failure to system log only
Before the creation of a new VM lease the engine verifies that
the lease doesn't already exist.
The verification is done by calling to GetVmLeaseInfoVDSCommand,
the command will fail in case the lease doesn't exist.

This patch prevents from logging the failure which is expected
to the engine audit log, the error log message can mislead the users
to think that the create lease operation failed.
Also, another log message added to the system log to indicate
that the lease is missing and the engine will create a new one.

Change-Id: I4b2da1cbb868b49f8942315056caaae8a4fc4af2
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>