core: Fix importing a VM with two disks from an OVA
Currently, when trying to import a VM from an OVA, and in case the VM
has more than one disk, the operation fails due to a
'duplicate key value' error.
(All VM's disks receive the same id, due to wrong use of the
generateNewDiskId(getVm().getImages(), disk) method).

After overcoming this issue, there was a yaml parsing error due to
using ', ' as a list's delimiter.

Those two issues were fixed in this patch.

Change-Id: Iecbdf72e91efc0ef2b96283651be8beec351022c
Bug-Url: https://bugzilla.redhat.com/1658249
Signed-off-by: Shani Leviim <sleviim@redhat.com>