core: Allow operations on HE VM/SD only to SU
This patch attempts to solve the issue when a user who is admin but not
superuser executes operations on HE VM/SD.

The patch allows such users to see the HE VM/SD but blocks operations on
those entities if the user is not a superuser.

This patch adds a new validate() method to VM/SD command base classes
that checks that the user is a superuser in case that the operation is
on HE VM/SD.
The relevant tests classes where modified accordingly and a sample test
that fails when accessing the operation on HE VM/SD with non-superuser
role were added to the update VM/SD test commands.

Change-Id: I3a3b7d3ab0d4f724b09b6a4f1f91ed2f5404ca60
Bug-Url: https://bugzilla.redhat.com/1511697
Signed-off-by: emesika <emesika@redhat.com>