ansible: Enabling VNC TLS does not add block markers
Ansible script for enabling VNC encryption (vnc_tls) did not add block
markers around the added content. Therefore, Ansible later was not able
to identify the addition and remove the content.

The block marker contained a reference to a variable which might not
have existed during host reinstall. This patch changes the marker to a
text with no dependencies.

Change-Id: I3f05326141a547e45a9cb154ae25d1f8e75674eb
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1676704
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>