spec: Require ovirt-vmconsole >= 1.0.7
Under certain conditions, upgrading ovirt-vmconsole from earlier
versions failed. So we do not want this to happen as part of
engine-setup, which will require a long and risky rollback. Make the
engine-setup package itself require it, so that if this fails, yum fails
updating the setup package, which is less dramatic.

Change-Id: Ideff77de419f9a5f8092f23a8e2acd0a71855d2f
Bug-Url: https://bugzilla.redhat.com/1678580
Signed-off-by: Yedidyah Bar David <didi@redhat.com>