webadmin: Fix wrong vnic profile state assumptions
The vNIC profile window was assuming that the previous state of
the check box for Port Morriring and Passthrough is correct.
Loading external network first caused that the checkboxes were
not editable even for valid configuration.

Moving call of the vNIC profile assignement check after the
external network check makes sure that actions are executed
in correct order.

Change-Id: I8bbc8d481e43584e81a3b9c6d11761875f573260
Bug-Url: https://bugzilla.redhat.com/1677426
Signed-off-by: Ales Musil <amusil@redhat.com>