core: Change IPv4 address converter into generic one
The converter used for cidr validation was capable of converting only
the IPv4 addresses into long. The new converter is capable of
converting both IPv4 and IPv6 into big int.

This conversion should be used for numerical manipulation with IP
addresses (and, or etc.). Because Java has internally all numbers
as signed. The IP representation can be stored as both
(signed/unsigned), but it must be consistent.

Change-Id: Iaeae2ce2c1a599d72941bc2828a2d858873fe599
Bug-Url: https://bugzilla.redhat.com/1670715
Signed-off-by: Ales Musil <amusil@redhat.com>