core: Error message for inaccessible glance
When trying to import an image from an inaccessible glance repository,
there's a RuntimeException with a full-trace.

This patch adds the following error to the engine log:
"Failed to import from the repository."
The same error appears on the UI and on the audit log.

In addition, while trying to import a non-exists image, an
OpenStackImageException is being thrown with a proper message.

Change-Id: Id05d1d7f0ad2fa6c439167da2625b28d98fe4cb7
Bug-Url: https://bugzilla.redhat.com/1530984
Signed-off-by: Shani Leviim <sleviim@redhat.com>