core: VNC encryption is always turned ON at upgrade
This is a resubmitting of the patch https://gerrit.ovirt.org/#/c/97582/,
which got overwritten due to a bad merge.

When the Ansible playbook is run to upgrade a host, host_deploy_vnc_tls
is not set, so it uses the default 'true' value, thus enabling
encryption on host regardless of VNC encryption setting for the Cluster.

The patch sets the Ansible variable at host upgrade.

Change-Id: Iec03381e4e2119ce2aa6c4df9b9c9a9d876ce6ad
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>