core: Add seclabel's to disks
libvirt dynamic ownership of devices was enabled in Vdsm 4.3.  File
and block based disks must be excluded from dynamic ownership because
disks are examined and handled also when the VM is not running.

To exclude a disk from dynamic ownership, a certain <seclabel> element
must be added to its domain XML definition.  This patch adds it.

Note that we still need to add <seclabel> elements on Vdsm side to
disks added on the host, such as replicas or snapshots.  But other
than that the functionality belongs to Engine, therefore this patch.

<seclabel> elements are meaningless but also harmless on Vdsm < 4.3,
where dynamic ownership is not enabled.

Change-Id: I23951c420f8f7f1a8f016b7d6b9f0f9ca29de6ec
Bug-Url: https://bugzilla.redhat.com/1666795
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>