engine: Update Deprecated CPU Types on Upgrade
When upgrading a Cluster to Compatibility Version 4.3
that includes Hosts that are not in maintenance mode and
when the CPU Type chosen was deprecated for said version,
the upgrade would fail because the previous CPU Type that
was deprecated could not be used to validate the
manufacturer.

This solution validates the previous CPU Type so that
only the manufacturer of valid CPU Types are
compared to the manufacturer of the current
CPU Type. Also when upgrading the Compatibility Version of
the Cluster when the CPU Type has been deprecated for the new
Compatibility Version, the CPU Type defaults to the
minimum CPU Type whose manufacturer is the same as the
manufacturer of the previous deprecated CPU Type.

Change-Id: I1499e5074576fdd0fb76aaad1c90b78a17f73f44
Bug-Url: https://bugzilla.redhat.com/1672859
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>