core: Export disks under vdsm user when exporting OVA
If VM disks are located on a NFS storage domain and the NFS directory is
exported with root_squash option, the disks are not readable by
processes running under root. In particular, this causes OVA export to
fail, because Ansible scripts used to build OVA are run under root.

This patch changes the scripts so as the part of the script that
accesses the disks is executed under vdsm user.

Change-Id: I10e25b1f448e3a4ce0678746e2ac63f91494324c
Bug-Url: https://bugzilla.redhat.com/1645229
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>