engine: Fixed Block Storage Failure
It appears that when the Destination Storage is a Blocked based
Storage Domain the libvirt sparseness does not fail, but continues and
sends an invalid length which breaks the truncation designed to remove
holes for sparseness support. However, allowing the libvirt to continue
by removing the truncation still causes a failure when the v2v vdsm
module attempts to process the file.

The fix is to  detect in the engine when the destination Storage Domain
is Block Storage and to set the Allocation Policy to Preallocated being
that Block Storage does not support Sparseness anyway.

Change-Id: I6c622c359ecdc82da45bacf7087f74dfe062c20b
Bug-Url: https://bugzilla.redhat.com/1663135
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>