engine: MaintenanceVds should be executed as a MaintenanceNumberOfVdss child
When executing MaintenanceNumberOfVdss, MaintenanceVds
command should be executed as a child of the command.
The status of MaintenanceNumberOfVdss should be set
to SUCCEEDED only after MaintenanceVdsCommand's status
is set to SUCCEEDED.

When executing host upgrade the HostUpgradeCallback
should check the status of MaintenanceNumberOfVdssCommand
before invoking host upgrade. Host upgrade should be
started only after MaintenanceNumberOfVdssCommand's
status is set to SUCCEEDED.

During MaintenanceVdsCommand if stopping gluster or VDO
service fails, the command should be marked as failed and
the Vds should be activated. This will restart the gluster
and VDO services and put the host back in UP status.

Change-Id: Idd9c7214bc49f21926bc646965ea46a5c1630333
Bug-Url: https://bugzilla.redhat.com/1679399
Signed-off-by: Ravi Nori <rnori@redhat.com>