Remove vmt of vm_interface
Previously, a vNIC had two references to a VM.
One was used to reference the VM if the VM is
a template, else the other reference was used.
For this reason only one reference was used
per vNIC, and the other reference was null.
This construct seems to confuse the constraint
checker of the database and enabled inconsistency.

This change merges both references into a single one,
which has to be set.

Bug-Url: https://bugzilla.redhat.com/1547038
Change-Id: I567a0fa81eb18255f3e9cf54491fdbe1a8966c93
Signed-off-by: Dominik Holler <dholler@redhat.com>
Signed-off-by: Eitan Raviv <eraviv@redhat.com>