webadmin: Removing confirmation when expanding a gluster volume
When adding more bricks to an existing gluster volume in a hc-setup
a popup confirmation was shown as "Expanding gluster volume in a
hyper-converged setup is not recommended as it could lead to degraded
performance. To expand storage for cluster,
it is advised to add additional gluster volumes.
Are you sure you want to continue?".
This patch removes the message

Change-Id: I01f38c0cdd77ed0a77f475885d1725ef97df2719
Bug-Url: https://bugzilla.redhat.com/1679109
Signed-off-by: Kaustav Majumder <kmajumde@redhat.com>