core: Explicitly set the MTU for an empty vnic profile
In cases where the MTU of a vnic is not specified 
(in the domain xml) on creation but later on when 
changing the backend network it changes to an explicit 
value, the operation will fail (requiring unplug/plug 
operation).
For the specific scenario of no-value to default-value 
to pass, on creation, the default value should be 
explicitly specified.

When the MTU is not specified, explicitly set it to 
the default of 1500.

This change keeps changing a plugged network to another
network with a different MTU failing because this might
lead to dropped packages on the vNIC.

Future patches will handle the domain xml in Engine 
and drop the handling in VDSM.

Change-Id: I13c5ccf16d4f8ee791d276f980daad5d890bcff5
Bug-Url: https://bugzilla.redhat.com/1676822
Signed-off-by: Ales Musil <amusil@redhat.com>
(cherry picked from commit 78c1c9ae4a5500f63d5e4886ca02eecf01a675e7)