restapi: Add upgrade action to cluster
Add /api/clusters/1234/upgrade action that sets
the upgrade_running flag of the cluster to true
or clears the upgrade_running flag for the cluster
based on the value of upgrade_action passed.

upgrade_action can take value of START or FINISH.

When an error occurs in setting the upgrade_running flag
an error is returned from the action.

Change-Id: Ic3d98f43f3d1253d8d8185fccb3e349a9f2c501a
Bug-Url: https://bugzilla.redhat.com/1663626
Signed-off-by: Ravi Nori <rnori@redhat.com>