Move from java 8 to java 11
Wildfly 15 supports java 11, its time to move on.
For list of java 11 features, comparison see dzone post series[1]

Major breaking changes with regards to dependencies:

- Unrelated EE modules removed from java
  - java.xml jaxb
  - javax.annotation
  - javax.activation

Those compile time dependencies where added to the relevant pom files.
They are shipped with wildfly anyway for runtime (those are EE deps we
are talking about)

maven-compiler-plugin for GWT remains unchanched, and will keep using java-1.8
when invoked, this keeps the GWT compliation exacly the same as it was.

This means every build system needs both java-11 and java-1.8 to compile
the complete product.

[1] dzone posts about java 11:
    - https://dzone.com/articles/90-new-features-and-apis-in-jdk-11
    - https://dzone.com/articles/90-new-features-and-apis-in-jdk-11-part-2

Bug-Url: https://bugzilla.redhat.com/1732738
Change-Id: I1a17efa56596b68e7258d3f48b7907a866840eec
Signed-off-by: Roy Golan <rgolan@redhat.com>