core: external refresh host with wait
Add wait for a refresh host that is triggered externally via the REST
API or the webadmin.

The wait guarantees that the request will be executed when it
succeeds in acquiring the relevant lock, rather than being aborted
immediately with a 'server is busy' message.

Until now, the wait policy was employed only in internal invocations of
the command to avoid collisions due to competing operations, but this
policy is now required also for external invocations due to the same
problem of competing operations - mainly instantiated in client-side
scripts.

In place of using a wait-forever policy as was previously the case for
internal invocations, a policy of wait-with-timeout is employed now in
order to protect JVM liveness.

The timeout is configurable via engine-config. The default value is 5
seconds.

The refresh host triggers a 'refresh host capabilities' command and
then a 'refresh host devices' command.

There is no state instantiation in the commands before the lock is
acquired so if a command needs to wait for the lock it is guaranteed
that it will use the latest available state when it is finally executed.

Change-Id: I79df73b49518504ee3773f318c787a4fb51057c9
Signed-off-by: Eitan Raviv <eraviv@redhat.com>