core: setup networks with timed wait
Add wait with timeout to setup networks command. The wait guarantees
that the request will be executed when it succeeds in acquiring the
relevant lock, rather than being aborted immediately with a 'server is
busy' message.

The timeout is configurable via engine-config. The default value is 5
seconds.

There is no state instantiation in the command before the lock is
acquired so if the command needs to wait for the lock it is guaranteed
that it will use the latest available state when it is finally executed.

Change-Id: Ie370a36f11e8b9074c3b6c1d2ebd21b30762ec3a
Signed-off-by: Eitan Raviv <eraviv@redhat.com>