backend+frontend: Support PassThroughCPU for all entities
Move useHostCpuFlags (Pass-Through Host CPU property) from VMStatic to
VMBase so that Templates, Snapshots and Pools will fully support it.

That fix also includes supporting useHostCpuFlags in all use cases including
High performance entities, import/export of a VM or a Template, REST api
flows etc.

In addition this fix includes supporting the useHostCpuFlags field value
when moving a VM/Template/Pool between clusters or when creating a VM based
on a template in a different cluster. This is done in the following manner:
If the target cluster still includes at least one assigned host then
useHostCpuFlags configuraiton is not changed, but if no hosts are included
in target cluster then useHostCpuFlags is disabled (unless it is a High
Performance entity).

Change-Id: I47443eac4ffbd70b159868d93beff7e4282b9e16
Signed-off-by: Sharon Gratch <sgratch@redhat.com>