core: CPU flags should be identical for migrating with CPUPassThrough
Fixing CpuLevelFilterPolicyUnit behavior so that in case of migrating a
VM with CPU Pass-Through enabled then the CPU flags configured for the
running VM (i.e. the CPU flags that the VM started running with) should
be exactly identical (and not just contained within) to the flags of the target host.

This is a QEMU requirement for the migration of a VM with CPU Pass
Through to succeed.

Change-Id: I942ebdb0d2f824f8eb2979f3538151ecf9e72879
Bug-Url: https://bugzilla.redhat.com/1686537
Signed-off-by: Sharon Gratch <sgratch@redhat.com>