core: A VM can be migrated together with others in positive affinity
The REST API call to migrate a VM takes an optional parameter,
'migrate_vms_in_affinity_closure'. When set,
any VMs that are in positive enforcing affinity with
the selected VM and run on the same host will be migrated
also.

Another change in this patch, is that GetValidHostsForVmsQuery
can consider the affinity closure too.

Change-Id: I35cb5e688e629290b233a24b83b14bf0964b383d
Bug-Url: https://bugzilla.redhat.com/1651406
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>