Host Certificate Reenrollment Done Too Late
Currently when upgrading a host, certification reenrollment
is done only if the old certificate of the host has already
expired. This patch causes reenrollment to be done if the
existing certificate is about to expire in the near future.
The time in the future is determined by the exiting configuration
variable which determines when a warning for eminent expiration
is given: CertExpirationAlertPeriodInDays

Bug-Url: https://bugzilla.redhat.com/1683281
Change-Id: Ida5ac4c89016afeed76873ab75f19a124f00b604
Signed-off-by: Ori_Liel <oliel@redhat.com>
(cherry picked from commit 252fdd38e184f2154b4f0ef7293cd3dc580cbf53)