core: Persist the guest session state into the DB
Session information provided by the vdsm like UserLoggedOn,
LoggedOff, Locked and so on is currently not persisted into
the database. This leads to vm_dynamic table update for every database
sync in VmAnalyzer. As soon as the session changes from Unknown
(initial state in the DB) to anything else the session field is
considered as differernt as the DB is not updated. This patche makes
the session state to be correctly stored into the DB.

Change-Id: I864c6de34ba72dba87ad450eea4b33921c357e5c
Bug-Url: https://bugzilla.redhat.com/1693628
Signed-off-by: Roman Hodain <rhodain@redhat.com>
(cherry picked from commit a003540301c52b9fd161f08d745feb3d59ee424e)