engine: make memory hotplug configurable
For RHEL 7, when a new memory is plugged in,
a special memory block of a fixed size is inserted first.
This allows unplug to work properly. However, this is
not necessary for other OS.

This patch makes the creation of a special memory block
configurable in osinfo by adding
new key devices.memoryHotplug.specialBlock.

Change-Id: I6991d8ea026bd74d23d4d70599e48ab91c7765ab
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>