core: allow cloning raw-sparse disk on file-based domain
Until now, when cloning a VM from a template,
the engine determined the volume type according to the
combination of the volume format and the storage domain type (file/block):

 - block based domain + raw format = preallocated volume type
 - block based domain + cow format = sparse volume type
 - file based domain + raw format = sparse volume type
 - file based domain + cow format = sparse volume type

After patch - https://gerrit.ovirt.org/#/c/93557/
was merged, VDSM supports copying/moving raw-preallocated file-based disks.

So now, the engine cannot determine the volume type according to the
combination of the volume format and the storage domain type (file/block),
the engine must refer to the user request.

Change-Id: I0dc04f6e9f8c4876b4ae8f896ff79d8df2f7ee73
Bug-Url: https://bugzilla.redhat.com/1696174
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>