webadmin: allow cloning raw-sparse disk on file-based domain
After patch https://gerrit.ovirt.org/#/c/99221/ will be merged,
the engine now doesn't determine the disks volume type according
to volume format and the storage domain combination
but according to the user selection.

In case of cloning a VM from a template using the UI, the user
can change only the disks formats (raw/cow), so the same behavior
 as creating a template from a VM will take place here:

 - cow  will be always sparse
 - raw + block-based domain = sparse
 - raw + file-based domain (CopyPreallocatedFileBasedDiskSupported=true) = source image volume type
 - raw + file-based domain (CopyPreallocatedFileBasedDiskSupported=false) = sparse

In order to support this change, we should have a VM object with
ClusterCompatibilityVersion initialized in the disk model.

Change-Id: Ie9835494e2e2ddfc62ed038cd092e1bb460095ea
Bug-Url: https://bugzilla.redhat.com/1696174
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>