core: add wait with timeout option to lock properties
Add a timeout configuration option to the LockProperties
configuration and streamline the API accordingly.

Currently any command using the locking mechanism with a wait, may
occupy resources (e.g. threads) indefinitely. Conversely any command
that does not use wait at all forces the request issuer to monitor the
failures and retry the request several times until it succeeds.

Both of these behaviors are dangerous for the liveliness of the JVM due
to increased usage of threads. Both of these behaviors also do not
guarantee any fairness in order of execution as requests are randomly
selected for execution (but randomness has better performance).

This has not been a problem with manual usage of the engine via the
web-admin but with more automatic tests being written for engine, a
growing number of requests collide due to the speed with which the
client issues them. These requests are revoked by engine in commands
that do not use wait, causing the tests to fail.
This in turn has brought about in recent years an increased usage of the
wait-forever facility in commands which endangers the liveliness of the
JVM in the long run.

Recently, setup networks and get capabilities requests have been used
extensively in tests and since they do not utilize any wait they were
revoked by engine when unable to acquire locks, causing tests to fail.
To handle these issues, this patch introduces a wait with timeout option
to the LockProperties configuration and streamlines the API accordingly.
The next patch introduces a wait with timeout functionality to the
LockManager.

It is intended that users of commands (e.g. tests) that have not been
using wait at all and would like the server to queue the commands for
execution instead, will be able to do so without endangering the
liveliness of the JVM by specifying a timeout for the wait.

Change-Id: I1e6f836a55d1ee118b1754d7103c5c9e62645ccc
Signed-off-by: Eitan Raviv <eraviv@redhat.com>