engine: Add VM Init when updating Cluster VMs
When updating the cluster such as when upgrading
the Compatibility Version, fetching and updating
the VMs of the cluster did not include the VM
Initial Run.

The fix was to add the fetching of the VM Initial
Run data within the Cluster Update code so that
the VM's Initial Run update function will
preserve the VM's Initial Run data.

Change-Id: I4e97a88d4b0c60c65ad04c95943fb1845929701f
Bug-Url: https://bugzilla.redhat.com/1699684
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>