core: Intern Strings holding application list
In a deployment with a large numbers of Windows VMs (>3k), the Engine
holds lists of VMs, each include an application list. The VMs on the
lists each contain the same, rather long String with the application
list. This put a pressure on JVM's heap, leading to OutOfMemory
condition in client's environment.

The underlying problem is in fact a design defect. This patch introduces
a quick-and-dirty fix to mitigate the RAM pressure.

Change-Id: I473cbf8b28888972b94f8efa3e89cedc4c860457
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1700725
Signed-off-by: Tomasz Baranski <tbaransk@redhat.com>