core: Avoid changing a DC's version with suspended VMs
Screenshot: https://imgur.com/a/cA0XLqA

When upgrading a DC's compatibility version, and that DC contains some
suspended VMs with an older custom compatibility version, those VMs
won't be able to resume once the upgrade has finished.
("The Custom Compatibility Version of VM X is not supported in Data
Center compatibility version y.y").

This patch adds a validation for that scenario and avoids it with an
error message.

Bug-Url: https://bugzilla.redhat.com/1710740
Change-Id: Iacae97705679d52613971368551ec93313479895
Signed-off-by: Shani Leviim <sleviim@redhat.com>