engine: VMWare UEFI import fix
When Importing from an external provider, the Bios Type
was not included in the import data which meant that
the Bios Type was not preserved and the i440fx/seabios
was used instead because this is the default.

For VMWare which has a post conversion processing and only
for libguestfs versions from 1.40, the Bios Type has been
added to the ovf file. This patch fetches the Bios Type from
the ovf data and updates the VM with the correct Bios Type.

The patch does not address the KVM which does not implement
Post Conversion nor older versions of libguestfs.

Change-Id: I4eb75f0a3ceaa1c375b26dd95681f9188cdebc0a
Bug-Url: https://bugzilla.redhat.com/1693571
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>