core: Persist hotplug via VmDevicesMonitoring
If a vNIC is activated on a running VM, VmDevicesMonitoring
is triggered to persist the updated device, instead of
persisting the device directly.

This prevents that the direct persisting is overwriting
data of VmDevicesMonitoring, which is responsible to keep the
db in sync with the devices of running VMs.

This might happen because the VmDevicesMonitoring might write to the
db before the vmDeviceDao.update(vmDevice) in updateDevice() is
executed. This timing is more likly to happen for SR-IOV vNICs, because
the SR-IOV flow in ActivateDeactivateVmNicCommand takes more time than
non-SR-IOV flows.
The VmDevicesMonitoring will not overwrite the incomplete information
written by vmDeviceDao.update(vmDevice) in updateDevice(), because this
write access is not detected by VmDevicesMonitoring, because
VmDevicesMonitoring updated the db after the last change of the hash
and VmDevicesMonitoring is assuming that it is the only writer to the
db.

Change-Id: I5873195053b2ca5ee7b20efa92870bb6f82a7e0e
Bug-Url: https://bugzilla.redhat.com/1703112
Signed-off-by: Dominik Holler <dholler@redhat.com>