core: SSH channel traffic during image conversion
When exporting a VM to OVA, disk image conversion process may take a
long time. Nothing is sent through the Ansible SSH channel during this
process, so the connection is closed by Ansible by timeout, while the
conversion script is still running and the disk is locked.

To prevent SSH channel to become idle, this patch switches on progress
notifications of qemu-img convert command. The only purpose of them is
to fill the channel with some traffic until the conversion is finished.

Change-Id: Id51960f6f7a909c767938a74970e6ac4e2650e1e
Bug-Url: https://bugzilla.redhat.com/1684266
Signed-off-by: Shmuel Melamud <smelamud@redhat.com>