engine: Remove default values in vm parameters
When VmManagementParametersBase is used for updating vm parameters, the
typical behaviour is that if the parameter value is null, its value should
not be changed. However, for consoleEnabled and affinityLabels there
were provided default values. So even if the user did not want to change
those values, they were changed to defaults (the console was disabled, no
affinity labels). This resulted in a bug when using the REST API.

Note that there were 2 more properties initialized in init() method -
graphicsDevices and storageDomainId but these were handled correctly in
commands:
- graphicsDevices - no action was taken if the map was empty
- storageDomainId - proper id is loaded at the command if needed

I've removed the init method completely and the properties are
initialized when declared. It is now the same as
VmTemplateManagementParameters.

Change-Id: I73389b83b60569367d632fe41d795b4a2ebb7eae
Bug-Url: https://bugzilla.redhat.com/1710696
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>