engine: Block Delete Protected Template change
When the Template is changed on a Pool where the Templates
are Delete Protected, the associated VMs are not deleted
nor are the new VMs recreated and various side affects
such as the inability to edit the Pool afterwards.

We now block the updating of a pool when the Template
has been changed and the VMs are Delete Protected so
that the use can only edit the template of the Pool when
the VMs are no longer Delete Protected.

Change-Id: If1ebb4062845f28833d2a27c7f57fecc4c06c0e3
Bug-Url: https://bugzilla.redhat.com/1659161
Signed-off-by: Steven Rosenberg <srosenbe@redhat.com>