restapi: update disk-attachment - fix null in href
When updating a disk-attachment under a VM, the href of the
disk-attachment returned contains the string 'null'. This is
the result of an exceptional case where the same entity
(disk-attachment) has the same parent (vm) in 2 different
locations in the API, causing ambiguity in the link generation
process. Enhancing the infrastructure to deal with such cases
in a generic way is complex and is not justified by this single
occurrence, hence the hardly-typed solution.

Bug-Url: http://bugzilla.redhat.com/1647018
Change-Id: I9c92ac05776ea6d2f853d0c56cc02948b98167e4
Signed-off-by: Ori_Liel <oliel@redhat.com>