webadmin: new representation icon for snapshot with memory
At virtual machines section, under snapshots tab,
snapshot with memory was represented with a chip
icon.

In this change, the chip icon has been changed,
and set to be a camera with a small memory icon,
at the upper right corner of the camera icon,
when hovering over the memory icon, a tooltip with
a title: "This snapshot contains memory" shows.

Screenshots - https://imgur.com/a/gUcBorX

Change-Id: I6f16d4a0ee48dfddcd7c936a3543b6353c3dc866
Bug-Url: https://bugzilla.redhat.com/1642266
Signed-off-by: Ahmad Khiet <akhiet@redhat.com>