core: Add labels to affinity groups
The affinity groups now have 2 new fields, "vmLabels" and "hostLabels".

When the affinity group is used for scheduling, the labels
are unpacked and the scheduler behaves as if all VMs from
labels in vmLabels and all hosts from labels in hostLabels
were added to the affinity group.

This patch also changes the behavior of lables.
Previously an affinity label behaved the same way
as a positive enforcing VM-to-host affinity group.
New labels do not act as affinity groups, unless the
field "implicitAffinityGroup" is set. This field is only
considered for cluster compatibility versions 4.3 and lower.

Change-Id: If6cd454312c84cfe352584f74eb48528ca0c0a47
Bug-Url: https://bugzilla.redhat.com/1680499
Signed-off-by: Andrej Krejcir <akrejcir@redhat.com>