webadmin: set Gluster disk default volume type to preallocated
A regression was introduced due to bugfix-
https://bugzilla.redhat.com/1644159
commit - 462ffbb391fcc0995154877b76b16a0cb5686eab

The fix caused setting the volume format of each Gluster based disk
to preallocated even if the user requested for sparse volume.

This patch reverts this behavior and set only the default value
of a Gluster based storage domain disk to be preallocated and allow to
the user to select different volume format.

Change-Id: I79a9bfbe240a31793334766d795da2322c78f8e2
Bug-Url: https://bugzilla.redhat.com/1704782
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>