engine: show vm's next run configuration
At the VMs table, a list of all changed fields
between current and next run configuration was added to the tooltip of
VM type column.

The list of changed fields is loaded on the server side. Since similar
logic has already been used in GetVmNextRunConfigurationQuery,
common parts were extracted to the VMHandler.

Change-Id: If331e058c5766f2d67a504c8907b0c44be5e37d1
Bug-Url: https://bugzilla.redhat.com/1711795
Signed-off-by: Lucia Jelinkova <ljelinko@redhat.com>