core: wait with timeout - satisfy coverity inspection
Coverity rejects the allocation of memory for an EngineLock object in a
local scope and then exiting this scope without calling the close()
method of this object. This happens when the lock could not be obtained.

This is because EngineLock implements the java.lang.AutoCloseable
interface, which is typically used when there is an underlying resource
that needs to be freed or closed using the close() method.
But EngineLock does not hold onto any resource, only some regular
internal memory members so Coverity's assumption is wrong.

To satisfy Coverity's inspection it would seem tempting to call
EngineLock#close() before leaving the scope when the lock cannot be
obtained but that would be counter productive.
EngineLock.close() calls LockManager#releaseLock() which in turn blocks
on the global lock of LockManager. So it would again try to access the
same locks that it has just timed-out on and which it does not hold,
incurring a performance penalty and achieving no result.

Therefore I am structuring the code in a manner consistent with that of
the CommandBase#acquireLockInternal() which satisfies Coverity's
inspection without any drawback to engine behavior.

Change-Id: Ib5db0b0cab1ea6ca728daa132aeb79835e381ed5
Signed-off-by: Eitan Raviv <eraviv@redhat.com>